requirements.txt
argcomplete==3.1.4
asgiref==3.7.2
aws-psycopg2==1.3.8
boto3==1.28.78
botocore==1.31.78
certifi==2023.7.22
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.5
django-cors-headers==4.2.0
django-ebhealthcheck==2.0.2
djangorestframework==3.14.0
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
placebo==0.9.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.1
typing_extensions==4.7.1
urllib3==1.26.18
Werkzeug==3.0.1
zappa==0.58.0
